+++ ArmPkg/Library/ArmExceptionLib/Arm/ArmException.c
@@ -11,7 +11,7 @@
 
 #include <Uefi.h>
 
-#include <Chipset/ArmV7.h>
+#include <Arm/AArch32.h>
 
 #include <Library/ArmLib.h>
