+++ ArmPkg/Library/ArmMmuLib/Arm/ArmMmuLibUpdate.c
@@ -17,7 +17,7 @@
 #include <Library/CacheMaintenanceLib.h>
 #include <Library/MemoryAllocationLib.h>
 
-#include <Chipset/ArmV7.h>
+#include <Arm/AArch32.h>
 
 #define __EFI_MEMORY_RWX  0                 // no restrictions
