+++ MdePkg/Include/Library/ArmLib.h
@@ -14,7 +14,7 @@
 #include <Uefi/UefiBaseType.h>
 
 #ifdef MDE_CPU_ARM
-  #include <Chipset/ArmV7.h>
+  #include <Arm/AArch32.h>
 #elif defined (MDE_CPU_AARCH64)
   #include <AArch64/AArch64.h>
 #else
